+++ paddle/fluid/inference/check_symbol.sh
@@ -18,7 +18,7 @@ lib=$1
 if [ $# -ne 1 ]; then echo "No input library"; exit -1 ; fi
 
 num_paddle_syms=$(nm -D "${lib}" | grep -c paddle )
-num_google_syms=$(nm -D "${lib}" | grep google | grep -v paddle | grep -v brpc | grep -c "T " )
+num_google_syms=$(nm -D "${lib}" | grep google | grep -v paddle | grep -v phi | grep -v brpc | grep -c "T " )
 
 if [ $num_paddle_syms -le 0 ]; then echo "Have no paddle symbols"; exit -1 ; fi
 if [ $num_google_syms -ge 1 ]; then echo "Have some google symbols"; exit -1 ; fi
