+++ paddle/fluid/platform/CMakeLists.txt
@@ -1,8 +1,5 @@
 proto_library(profiler_proto SRCS profiler.proto DEPS framework_proto
               simple_threadpool)
-if(WITH_GPU)
-  proto_library(external_error_proto SRCS external_error.proto)
-endif()
 if(WITH_PYTHON)
   py_proto_compile(profiler_py_proto SRCS profiler.proto)
   add_custom_target(profiler_py_proto_init ALL COMMAND ${CMAKE_COMMAND} -E
