+++ paddle/phi/backends/callback_manager.cc
@@ -17,8 +17,8 @@
 #include <ThreadPool.h>
 
 #include "paddle/fluid/platform/device/device_wrapper.h"
-#include "paddle/fluid/platform/enforce.h"
 #include "paddle/phi/backends/device_guard.h"
+#include "paddle/phi/core/enforce.h"
 
 namespace phi {
