+++ paddle/phi/backends/gpu/gpu_resources.cc
@@ -32,8 +32,7 @@
 
 #include "unsupported/Eigen/CXX11/Tensor"
 
-// TODO(phi): remove fluid header.
-#include "paddle/fluid/platform/enforce.h"
+#include "paddle/phi/core/enforce.h"
 
 namespace phi {
