+++ paddle/phi/kernels/funcs/concat_funcs.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "paddle/fluid/platform/enforce.h"
+#include "paddle/phi/core/enforce.h"
 #include "paddle/phi/core/errors.h"
 namespace phi {
 namespace funcs {
