+++ paddle/phi/kernels/funcs/gru_compute.h
@@ -12,7 +12,7 @@ limitations under the License. */
 #pragma once
 
 #include "paddle/fluid/platform/device_context.h"
-#include "paddle/fluid/platform/enforce.h"
+#include "paddle/phi/core/enforce.h"
 #include "paddle/phi/kernels/funcs/detail/activation_functions.h"
 
 namespace phi {
