+++ paddle/phi/kernels/gpu/class_center_sample_kernel.cu
@@ -30,7 +30,7 @@ namespace cub = hipcub;
 #include <random>
 
 #include "paddle/fluid/framework/tensor_util.h"
-#include "paddle/fluid/platform/enforce.h"
+#include "paddle/phi/core/enforce.h"
 
 #if defined(PADDLE_WITH_NCCL) || defined(PADDLE_WITH_RCCL)
 #include "paddle/fluid/distributed/collective/ProcessGroup.h"
